Part1/2 Autoggrad.Function Impl for attn
